sorting and many small adjustments made

+++ src/components/film/FilmPage.jsx
@@ -1,49 +1,63 @@
-import React, { useEffect, useState } from 'react'
-import '../../styles/film-page.css';
 import axios from 'axios';
+import React, { useEffect, useState } from 'react';
 import { NavLink, useNavigate, useParams } from 'react-router-dom';
-import Cast from './Cast';
-import Review from './Review';
-import Base from '../main/Base';
-import { getUser } from '../services/user-service';
-import { getCurrentUser } from '../auth/auth';
-import { Button } from 'reactstrap';
-import { deleteFilm } from '../services/film-service';
 import { toast } from 'react-toastify';
+import { Button } from 'reactstrap';
+import '../../styles/film-page.css';
+import { getCurrentUser } from '../auth/auth';
+import Base from '../main/Base';
+import { deleteFilm } from '../services/film-service';
+import { getUser } from '../services/user-service';
+import Cast from './Cast';
+import FilmCard from './FilmCard';
+import Review from './Review';
 
 const FilmPage = () => {
   let params = useParams();
   const [film, setFilm] = useState({});
+  const [films, setFilms] = useState([]);
   const [reviews, setReviews] = useState([]);
   // const [currentUser, setCurrentUser] = useState({});
   const [userRole, setUserRole] = useState([]);
 
+  const [filmCategory, setFilmCategory] = useState(1);
+  
   useEffect(()=>{
     getUser(getCurrentUser()).then((response)=>{
-      console.log(response);
+      // console.log(response);
       setUserRole(response?.roles);
       // console.log(userRole);
     }).catch((error)=>{
     })
     axios.get(`http://localhost:8080/films/read/${params.id}`)
     .then((response)=>{
-      console.log(response);
+      // console.log(response);
       setFilm(response.data);
       setReviews(response.data.reviews);
+      setFilmCategory(response.data.category?.categoryId);
     })
-  }, [params.id])
+    axios.get(`http://localhost:8080/films/read/category/${filmCategory}`)
+    .then((response)=>{
+      // console.log(response);
+      setFilms(response.data);
+    }).catch((error)=>{
+      console.log(error);
+    })
+  }, [params.id, filmCategory])
 
   const navigate = useNavigate();
 
   const handleDelete = (e)=>{
-    e.preventDefault();
-    deleteFilm(params.id).then((response)=>{
-      console.log(response);
-      navigate('/');
-      toast.success("Film deleted successfully");
-    }).catch((error)=>{
-      console.log(error);
-    })
+    if (window.confirm(`Are you sure you want to delete this film?`)) {
+      e.preventDefault();
+      deleteFilm(params.id).then((response)=>{
+        // console.log(response);
+        navigate('/');
+        toast.success("Film deleted successfully");
+      }).catch((error)=>{
+        console.log(error);
+      })
+    }
   }
 
 
@@ -118,13 +132,34 @@ const FilmPage = () => {
             <iframe width='700px' title='trailer'
               height='400px' 
               src={film.trailerLink}
-              frameBorder="0">
+            >
             </iframe>
           </span>
         )
       }
       <Cast title={film.title} casts={film.cast}/>
       <Review id={params.id} title={film.title} reviews={reviews}/>
+      <div>
+        {/* <div>You may also like</div> */}
+        
+        <div style={{margin: '100px', marginTop: '-10px'}}>
+          <div style={{color: 'black', margin: '30px 0px -30px 30px', fontSize: '25px'}}>
+          You may also like
+          </div>
+          <div className="film-grid">
+            {films.slice(0, 5).map((film, i) => (
+              String(film.id) !== String(params.id) && ( // this could cause a bug
+                <div key={i} onClick={()=>{
+                  navigate(`/films/${film.id}`)
+                  window.scrollTo(0, 0)
+                }}>
+                  <FilmCard {...film} />
+                </div>
+              )
+            ))}
+          </div>
+        </div>
+      </div>
     </Base>
   )
 }
